Introduce first version of gap iteration functionality
This change introduces the first version of the gap iteration
functionality. We provide a trait, Gappable, that is currently
implemented for all iterators that produce references to the iterated
values. This trait in turns provides a method, gaps, that converts the
underlying iterator into one that resolves to all gaps in a certain
range of values.