Introduce RangeGappable trait to cover collections with range() method
Certain collections provide the means for iterating over only a subset
of the contained values (e.g., BTreeSet), represented by a range. It
makes sense for us to treat these types a bit specially, as that may
lead to improved gap iteration performance.
With this change we introduce an additional trait, RangeGappable, that
adds a gaps() method to these types directly, without requiring the
explicit use of an iterator. This functionality is mostly provided in an
attempt to reduce boiler plate in client code and reducing the risk of
working with the "wrong" type of iterator accidentally.