Export bounds function for retrieving bounds of a range
It can be very painful to work with a range because it cannot easily be
cloned. The most likely work flow is to extract the start and end
bounds, but even that is not trivially possible because the extracted
bounds contain reference values.
To prevent client code from reinventing the wheel all over again, this
change exports the bounds function that accomplishes this job for inner
types that are copyable.