Change GapIter to assume non-reference types in wrapped iterator
Due to current restrictions surrounding Rust's Bound type, we require
that all our types of which ranges are comprised be copyable. That in
turn means that the objects over which we iterate also be copyable,
because these types are directly related.
The Copy trait (which marks types as copyable) is only implemented for
types for which it is cheaper to just copy them as opposed to working
with references. As such, our current implementation of GapIter, which
assumes that the wrapped iterator yields references to items that by
itself have to be copyable, is ultimately wrongly forcing people to use
iterators over references.
This change fixes this problem, by removing the reference indirection
from the GapIter struct.