Fix release build compilation failure
The debug_assert macro type checks the arguments passed to it,
irrespectively of whether debug assertions are enabled or not. On a
build with debug assertions disabled, that would cause compilation of
the crate to fail because the self.last member of the GapIter struct was
not present.
This change fixes the issue by moving the assertion into the compile
time conditioned block. To prevent such issues from happening in the
future, we make sure to compile release builds in the CI pipeline.
Lastly, while at it, update the cargo invocation line in the CI script.