Remove Copied type from RangeGappable::gaps return type
This change adjusts the signature of the RangeGappable::gaps method to
return GapIter<Self::Iter, V> instead of GapIter<Copied<Self::Iter>, V>.
That is, we push the Copied type of the return type into the associated
Iter type. In so doing we provide more flexibility to implementors of
the trait, which now no longer have to use Copied in the return type of
this method.