Fix "implementation of FnOnce is not general enough" error
This change fixes a compile error that we saw on some client sites.
Instead of relying on the Copied iterator type in the RangeGappable
implementation for BTreeMap we adjust our mapping function to eliminate
the passed in reference through a copy directly.
The exactly reason for this problem is not clear, but it has to do with
the compiler not being able to infer types properly in some
circumstances.