Remove overly strong assertion
This change removes the debug assertion checking that the user did not
provide a range that has its end bound before the start bound to
GapIter::new.
We had this assertion in place to catch unintentional usage. However,
there are legitimate cases where it is just too cumbersome to special
case "invalid" ranges (as per this definition). What's more, at least
for BTreeSet and BTreeMap there are already checks in place that cause
panic when the end bound is lower than the start. However, while we
flagged cases such as 2..2 (that's (Included(2), Excluded(2))), the
standard library actually allows them.